requirements.txt
streamlit-searchbox==0.1.3